rename incompleteManifest to manifest
